requirements.txt
fastapi==0.95.2 
uvicorn[standard]==0.23.2 
requests==2.31.0 
numpy==1.24.4 
pandas==1.5.3 
python-dotenv==1.0.0 
openai==1.54.0 
